Start/stop SnapclientService via broadcast intents
